리팩토링(첫번째) - import 부분에서 npm 모듈과 src 폴더 파일 분리 - index.js에서 app.js 분리 - button의 children 변수 제거
